+++ migrations/2026-06-29-020000_order_cancelled_status/down.sql
@@ -1,0 +1,16 @@
+-- This file should undo anything in `up.sql`
+ALTER TABLE orders
+    ALTER COLUMN status
+    TYPE text
+    USING status::text;
+
+DROP TYPE dborderstatus;
+
+CREATE TYPE dborderstatus AS ENUM (
+    'New', 'InProgress', 'ReadyToShip', 'Shipping', 'Done'
+);
+
+ALTER TABLE orders
+    ALTER COLUMN status
+    TYPE dborderstatus
+    USING status::dborderstatus;