+++ migrations/2026-06-29-030000_order_payment_status/down.sql
@@ -1,0 +1,5 @@
+-- This file should undo anything in `up.sql`
+ALTER TABLE "orders"
+DROP COLUMN payment_status;
+
+DROP TYPE dbpaymentstatus;