+++ migrations/2026-06-29-030000_order_payment_status/up.sql
@@ -1,0 +1,7 @@
+-- Your SQL goes here
+CREATE TYPE dbpaymentstatus AS ENUM (
+    'Pending', 'Paid', 'Failed', 'Refunded'
+);
+
+ALTER TABLE "orders"
+ADD COLUMN payment_status dbpaymentstatus NOT NULL DEFAULT 'Pending';